(chore) Remove redundant signal handler flags
They were used but could be expressed in the existing flags and thus
were redundant.

This has the side effect of printing when piped commands are finished as
well. Given the constraints of the assignment and the fact that pipes
weren't part of the specification, this is deemed to be ok as it doesn't
hinder any functionality.